Refactor 1.2
Series of Refactor 2

Changed game to a singleton class, but all of it's members (except gamestate) were made private. This enables the game to be a connecting class, that allows other core systems to interact with eachother, but not be dependent eachother. This allows the use of a singleton without worrying about wild dependencies.