Refactor 1.3
Decided against the heavy reliance on the game manager. Decoupled classes that could be decoupled, and for now, passing in variables explicitly that need using the game manager as a sort of dependency injector/manager. Allows the game manager to init everything, but also let's dependencies be more visible since it's not all passing through the game manager anymore.