Random Encounters
Random Encounters now working via the region system using the player.

Added an encounter rate for the party members to create a cumilitave party encounter rate that is got by the player as it is moving around.

Random encounters send region data to battle manager and generate battles.

If a map doesn't have a region, it's simply a no encounter area. ( You can also set a region to have no encounters, but I am unsure whether I want this or not).