Battle Turn Order
Turn order system that loops and is connected to turnbar UI

Uses a priority queue to get the next battle actor in the list, constantly keeping a list of 6 to keep the UI updated.

The current battle speed is not balanced, but the mechanics are there, and abilities like haste and slow should be implementable pretty easily.